Move malloc_aligned.hpp out of supernova and into common
`server/supernova/utilities/malloc_aligned.hpp` is part of supernova,
but both sclang and scsynth include it. This commit moves the file to
`common/` to better reflect its dependencies.

Both `include/common/` or `common/` are possible locations for this
file, but to my understanding `include/common/` is for APIs exposed to
third parties, which we probably don't want here.

This fixes #3362.
