+++ server/supernova/server/audio_bus_manager.hpp
@@ -22,7 +22,7 @@
 #include <cstdint>
 #include "sample_types.hpp"
 
-#include "../utilities/malloc_aligned.hpp"
+#include "malloc_aligned.hpp"
 #include "nova-tt/rw_spinlock.hpp"
 
 namespace nova {
